+++ src/Symfony/Component/Mime/Tests/MessageTest.php
@@ -245,7 +245,7 @@ EOF;
 EOF;
 
         $extractor = new PhpDocExtractor();
-        $propertyNormalizer = new PropertyNormalizer(null, null, $extractor);
+        $propertyNormalizer = new PropertyNormalizer(null, null, $extractor, null, null, [], true);
         $serializer = new Serializer([
             new ArrayDenormalizer(),
             new MimeMessageNormalizer($propertyNormalizer),