+++ src/Symfony/Component/Serializer/Normalizer/GetSetMethodNormalizer.php
@@ -54,6 +54,10 @@ class GetSetMethodNormalizer extends AbstractObjectNormalizer
     ) {
         parent::__construct($classMetadataFactory, $nameConverter, $propertyTypeExtractor, $classDiscriminatorResolver, $objectClassResolver, $defaultContext);
         $this->allowNormalizationOfObjectsWithoutAnyGetters = $allowNormalizationOfObjectsWithoutAnyGetters;
+
+        if (\func_num_args() < 7) {
+            trigger_deprecation('symfony/serializer', '6.2', '$allowNormalizationOfObjectsWithoutAnyGetters parameter of %s() should be explicitly provided since the default value will change to `true` in symfony/serializer >=7.0', __METHOD__);
+        }
     }
 
     /**